feat(bb merge): during pr merge, if delete-source-branch is set to True bb will checkout & delete
Upon pr merge, if delete-source-branch is set to true bb will now checkout to target branch post pr
merge and shall delete the local ref of the from branch in addition to deleting the branch in
remote. Upon successful deletion it fetches the latest remote changes to the local workspace.